backend/requirements.txt
# Backend dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.13.0

# ClickHouse
clickhouse-driver>=0.2.6
aiochclient>=0.2.0

# Redis
redis>=5.0.0
aioredis>=2.0.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# AI SDKs
anthropic>=0.18.0
openai>=1.12.0
langchain>=0.1.0
langchain-anthropic>=0.1.0
langchain-openai>=0.0.5

# Async task processing
celery>=5.3.0
celery-redbeat>=2.2.0

# Data processing
pandas>=2.1.0
numpy>=1.26.0
polars>=0.20.0  # Faster alternative for large files

# Log parsing
grok>=1.0.0
regex>=2023.10.0
python-dateutil>=2.8.0

# Email
aiosmtplib>=3.0.0
email-validator>=2.1.0

# Report generation
jinja2>=3.1.0
markdown>=3.5.0
weasyprint>=60.0  # PDF generation

# Security
cryptography>=42.0.0
python-dotenv>=1.0.0

# HTTP client
httpx>=0.26.0
aiohttp>=3.9.0

# Monitoring
prometheus-client>=0.19.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
httpx>=0.26.0  # For testing FastAPI

# Utilities
orjson>=3.9.0  # Fast JSON
tenacity>=8.2.0  # Retry logic
structlog>=24.1.0  # Structured logging